web/backend/requirements.txt
# Core dependencies from original nano-dlna
tqdm>=4.65.0
ffmpeg-python>=0.2.0
Twisted>=23.10.0
setuptools>=69.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Web application dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
jinja2>=3.1.2
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=11.0.0

# Development dependencies
pytest>=7.4.0
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Depth processing dependencies - commented out to make the app start without these
# If you need depth processing, uncomment and install these manually:
# numpy>=1.20.0
# opencv-python>=4.5.0
# scikit-learn>=1.0.0

# Image processing dependencies (required for projection animation)
pillow>=9.0.0

streamlit>=1.10.0
